GameSimulator -> BlackJackGame, Participant 객체 파일 생성, OperationWithCard.java 생성
